Move 'get_disabled_container_list' to dut_utils.py (#3335)
Description of PR
Summary:
Util get_disabled_container_list is duplicated defined in both test_container_autorestart and test_container_check.
This PR move get_disabled_container_list definiton to dut_utils.py, and fix always_disabled issue.

Approach
What is the motivation for this PR?
This PR is to move get_disabled_container_list definiton to dut_utils.py, and fix always_disabled issue.

How did you do it?
Please see code.

How did you verify/test it?
Verified on SN4600c.
